+++ source/Delphi.Forms.Info.pas
@@ -3,7 +3,7 @@ unit Delphi.Forms.Info;
 interface
 
 const
-  Version = '0.2.15.0';
+  Version = '0.2.16.0';
 
 implementation